backend/requirements.txt
Django==5.2.4
google_api_python_client==2.177.0
google_auth_oauthlib==1.2.2
openai==1.98.0
protobuf==6.31.1
python-dotenv==1.1.1
Requests==2.32.4
